arm64: dts: qcom: msm8916-samsung-j3ltetw: Add touchscreen
J3 and J5X use an Imagis IST3038C touchscreen that is connected to
blsp_i2c5. Add it to the device tree.

Regulator and MFD drivers for SM5703 are unavailable at the moment, which
is required to power up the touchscreen on J5, so it's disabled on J5 for
now.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>