wifi: ath10k: add support for QCA9379 hw1.0 SDIO
Add support for the QCA9379 chipset. Analogous to the vendor qcacld driver
this change treats the QCA9379 family the same as the QCA9377 family.

The card generally works and connects to a variety of networks (open,
WPA2-PSK, WPA3 Personal) but currently does not roam cleanly and will
reauthenticate, causing intermittent loss of connectivity.

This change was tested on a device that uses a LITE-ON WCBN3510A card.

Tested-on: QCA9379 hw1.0 SDIO WLAN.NPL.1.6-00163-QCANPLSWPZ-1

Signed-off-by: Felix Kaechele <felix@kaechele.ca>